Init tasks after toothpick has been configured
